+++ exporter/exporterhelper/tracehelper.go
@@ -93,7 +93,11 @@ func NewTraceExporter(
 	be := newBaseExporter(cfg, logger, options...)
 	be.wrapConsumerSender(func(nextSender requestSender) requestSender {
 		return &tracesExporterWithObservability{
-			obsrep:     obsreport.NewExporter(configtelemetry.GetMetricsLevelFlagValue(), cfg.Name()),
+			obsrep: obsreport.NewExporter(
+				obsreport.ExporterSettings{
+					Level:        configtelemetry.GetMetricsLevelFlagValue(),
+					ExporterName: cfg.Name(),
+				}),
 			nextSender: nextSender,
 		}
 	})