+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## 🛑 Breaking changes 🛑
 
+- Rename pdata.DoubleSummary to pdata.Summary (#2774)
 - Refactored `consumererror` package (#2768)
   - Eliminated `PartialError` type in favor of signal-specific types
   - Renamed `CombineErrors` to `Combine`